demand-forecasting-api/requirements.txt
fastapi
uvicorn
pandas
prophet
supabase
python-dotenv 